+++ app/src/androidTest/java/com/forrestguice/suntimeswidget/themes/SuntimesThemeTest.java
@@ -29,6 +29,7 @@ import com.forrestguice.suntimeswidget.R;
 import com.forrestguice.suntimeswidget.SuntimesActivity;
 import com.forrestguice.suntimeswidget.SuntimesActivityTestBase;
 import com.forrestguice.suntimeswidget.settings.WidgetThemes;
+import com.forrestguice.suntimeswidget.themes.defaults.DarkTheme;
 
 import org.junit.Rule;
 import org.junit.Test;
