+++ app/src/main/java/com/forrestguice/suntimeswidget/themes/ImportThemesTask.java
@@ -25,7 +25,6 @@ import android.support.annotation.Nullable;
 import android.util.Log;
 
 import java.io.BufferedInputStream;
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
